Fixed rocket.acc.applyForce(forcex,forcey)
Instead of adding an x and y componet to the exiting accerlation vector,
I decided to just create a brand new vector with a forceX and a forceY
direction. This may cause performance issues down the line, but I'll
keep a track on it and do some more work as to why adding an x and y
componet is not working.